Refactor for Test Breakout, removed css
