--- src/app/rating/Rating.module.css
@@ -1,18 +1,0 @@
-/* Rating.module.css */
-.rectangle {
-    width: 300px;
-    height: 100px;
-    background-color: #f0f0f0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 1px solid #ccc;
-    border-radius: 8px;
-    margin-top: 20px;
-}
-
-.rectangle p {
-    margin: 0;
-    font-size: 16px;
-    color: #333;
-}
